fix store structure
