+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [0.2.10] - 2020-09-24
+## [0.2.11] - 2020-09-24
 * Support for custom disk storage.
 
 ## [0.2.9] - 2020-09-15
