+++ lib/dio_http_cache.dart
@@ -2,4 +2,6 @@ library dio_http_cache;
 
 export 'src/builder_dio.dart';
 export 'src/core/config.dart';
+export 'src/core/obj.dart';
 export 'src/manager_dio.dart';
+export 'src/store/store_impl.dart';
