+++ lib/src/core/manager.dart
@@ -3,13 +3,15 @@ import 'dart:convert';
 import 'package:crypto/crypto.dart';
 import 'package:dio_http_cache/src/core/config.dart';
 import 'package:dio_http_cache/src/core/obj.dart';
-import 'package:dio_http_cache/src/core/store.dart';
+import 'package:dio_http_cache/src/store/store_disk.dart';
+import 'package:dio_http_cache/src/store/store_impl.dart';
+import 'package:dio_http_cache/src/store/store_memory.dart';
 import 'package:sqflite/utils/utils.dart';
 
 class CacheManager {
   CacheConfig _config;
-  BaseCacheStore _diskCacheStore;
-  BaseCacheStore _memoryCacheStore;
+  ICacheStore _diskCacheStore;
+  ICacheStore _memoryCacheStore;
   MD5 _md5;
   Utf8Encoder _utf8encoder;
